server/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.2
gunicorn==21.2.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==0.21.0
psycopg2==2.9.3
# psycopg2-binary==2.9.3
Werkzeug==2.3.6